Small fix.
git-svn-id: https://svn.dealii.org/trunk@8803 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/bits/vector_19.cc
@@ -28,7 +28,7 @@ void test (Vector<double> &v)
   for (unsigned int i=0; i<v.size(); i+=1+i)
     {
       v(i) = i;
-      norm = std::max(norm,fabs(i));
+      norm = std::max(norm,std::fabs(static_cast<double>(i)));
     }
   v.compress ();
